separate commander and inquirer logic
